+++ src/local_lib/headers/node_transactor.h
@@ -4,7 +4,7 @@
 
 #include "transaction_list.h"
 
-typedef int(*Reward)(struct node *self, int percentage, Bool use_default, float * block_tot_reward);
+typedef int(*Reward)(struct node *self, float percentage, Bool use_default, float * block_tot_reward);
 struct node {
     pid_t pid;
     int node_id; /* Index id of the node into the message queues for nodes*/
@@ -12,7 +12,7 @@ struct node {
     float budget;
     int tp_size;
     int block_size;
-    int percentage;
+    float percentage;
     Queue transactions_pool;/*Make it a pointer to shm*/
     Queue transactions_block;
     Reward calc_reward;
@@ -27,7 +27,7 @@ struct node {
  * @param block_size size of the transaction block
  * @param calc_reward pointer to function that calculate the reward
  */
-void node_create(struct node *self, pid_t node_pid, int node_id, float budget, int tp_size, int block_size, int percentage,Reward calc_reward);
+void node_create(struct node *self, pid_t node_pid, int node_id, float budget, int tp_size, int block_size, float percentage,Reward calc_reward);
 
 /**
  * Free the memory from the queue list of transactions
@@ -51,5 +51,5 @@ int update_budget(struct node *self);
  * @param block_tot_reward the total block reward value
  * @return -1 in case of FAILURE. 0 otherwise
  */
-int calc_reward(struct node* self, int percentage, Bool use_default, float * block_tot_reward);
+int calc_reward(struct node* self, float percentage, Bool use_default, float * block_tot_reward);
 #endif /*DGT_SO_PROJECT_21_22_NODE_TRANSACTOR_H*/
